Fixup some more process.env problems
