+++ package.json
@@ -2308,7 +2308,9 @@
         "node-has-native-dependencies": "^1.0.2",
         "node-html-parser": "^1.1.13",
         "node-loader": "^2.0.0",
+        "node-polyfill-webpack-plugin": "^1.1.4",
         "nyc": "^15.0.0",
+        "path-browserify": "^1.0.1",
         "postcss": "^7.0.36",
         "postcss-cssnext": "^3.1.0",
         "postcss-import": "^12.0.1",
